NSPV-238 feat(entity): atualiza as anotações de validação e melhora os relacionamentos das entidades

+++ src/main/java/com/senai/npsv_gestor_tintas_backend/application/service/ItemFormulaService.java
@@ -6,6 +6,7 @@ import com.senai.npsv_gestor_tintas_backend.domain.entity.ItemFormula;
 import com.senai.npsv_gestor_tintas_backend.domain.repository.FormulaRepository;
 import com.senai.npsv_gestor_tintas_backend.domain.repository.ItemFormulaRepository;
 import com.senai.npsv_gestor_tintas_backend.domain.repository.ProdutoRepository;
+import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -15,6 +16,7 @@ import java.util.List;
 @Service
 @RequiredArgsConstructor
 public class ItemFormulaService {
+
     private final ItemFormulaRepository itemRepository;
     private final FormulaRepository formulaRepository;
     private final ProdutoRepository produtoRepository;
@@ -22,17 +24,35 @@ public class ItemFormulaService {
     @Transactional
     public ItemFormulaResponseDTO adicionarInsumoNaReceitaDaFormula(ItemFormulaRequestDTO dto) {
         ItemFormula item = dto.toEntity();
-        item.setFormula(formulaRepository.findById(dto.formulaId()).orElseThrow(() -> new RuntimeException("Fórmula não encontrada.")));
-        item.setInsumo(produtoRepository.findById(dto.insumoId()).orElseThrow(() -> new RuntimeException("Insumo não encontrado no estoque.")));
-        return ItemFormulaResponseDTO.fromEntity(itemRepository.save(item));
+
+        item.setFormula(
+                formulaRepository.findById(dto.formulaId())
+                        .orElseThrow(() -> new EntityNotFoundException("Fórmula não encontrada."))
+        );
+
+        item.setInsumo(
+                produtoRepository.findById(dto.insumoId())
+                        .orElseThrow(() -> new EntityNotFoundException("Insumo não encontrado no estoque."))
+        );
+
+        ItemFormula itemSalvo = itemRepository.save(item);
+        return ItemFormulaResponseDTO.fromEntity(itemSalvo);
     }
 
+    @Transactional(readOnly = true)
     public List<ItemFormulaResponseDTO> consultarInsumosDaReceita(String formulaId) {
-        return itemRepository.findByFormulaId(formulaId).stream().map(ItemFormulaResponseDTO::fromEntity).toList();
+        return itemRepository.findByFormulaId(formulaId)
+                .stream()
+                .map(ItemFormulaResponseDTO::fromEntity)
+                .toList();
     }
 
     @Transactional
     public void removerInsumoDaReceita(String id) {
+        if (!itemRepository.existsById(id)) {
+            throw new EntityNotFoundException("Item da fórmula não encontrado.");
+        }
+
         itemRepository.deleteById(id);
     }
 }
